airline_recommends_20250429/_apps/requirements.txt
dash==2.18.1
dash-bootstrap-components==1.6.0  # システム標準バージョン
databricks-sql-connector>=3.4.0   # 必須（デフォルトで入っているが明示的に指定）
pandas>=2.0.0                     # データ処理用（デフォルト未インストールの可能性）
databricks-sdk>=0.33.0            # Databricks連携用
mlflow-skinny>=2.16.2             # モデル管理用（軽量版）
gunicorn==23.0.0                  # 本番環境用WSGIサーバー
qrcode